# swagger.yaml
openapi: 3.0.1
info:
  title: Тестовый проект Медузы — API
  version: 1.0.0
servers:
  - url: http://mac.mini:3000/
    description: "Pavel Manylov development server"
  - url: http://localhost:3000
    description: "localhost Server"
tags:
  - name: posts
    description: Публикации
  - name: components
    description: Компоненты
paths:
  /posts:
    post:
      tags:
        - posts
      summary: Создать новую публикацию
      operationId: createPost
      requestBody:
        description: Публикация, которую необходимо добавить
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPost'
        required: true
      responses:
        200:
          description: Созданная публикация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        422:
          description: Невалидные данные
          content: { }
        500:
          description: Внутренние ошибки
          content: { }
      x-codegen-request-body-name: body

  /posts/{post_id}:
    get:
      tags:
        - posts
      summary: Получить публикацию по идентификатору
      operationId: getPost
      parameters:
        - name: post_id
          in: path
          description: Идентификатор публикации
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Запрошенная публикация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        422:
          description: Невалидные данные
          content: { }
        500:
          description: Внутренние ошибки
          content: { }
      x-codegen-request-body-name: body
    put:
      tags:
        - posts
      summary: Обновление публикации
      operationId: updatePost
      requestBody:
        description: Актуальные сведения о публикации
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
        required: true
      parameters:
        - name: post_id
          in: path
          description: Идентификатор публикации
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Обновлённая публикация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        422:
          description: Невалидные данные
          content: { }
        500:
          description: Внутренние ошибки
          content: { }
      x-codegen-request-body-name: body
  /posts/{post_id}/components:
    get:
      tags:
        - components
      summary: Получить все компоненты данной публикации
      operationId: indexComponents
      parameters:
        - name: post_id
          in: path
          description: Идентификатор публикации
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Список компонентов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Component'
        422:
          description: Невалидные данные
          content: { }
        500:
          description: Внутренние ошибки
          content: { }
      x-codegen-request-body-name: body
    post:
      tags:
        - components
      summary: Создать новый компонент в публикации
      operationId: createComponent
      requestBody:
        description: Компонент, который необходимо добавить
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewComponent'
        required: true
      parameters:
        - name: post_id
          in: path
          description: Идентификатор публикации
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Созданный компонент
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Component'
        422:
          description: Невалидные данные
          content: { }
        500:
          description: Внутренние ошибки
          content: { }
      x-codegen-request-body-name: body
  /posts/{post_id}/components/{component_id}:
    get:
      tags:
        - components
      summary: Сведения о компоненте
      operationId: getComponent
      parameters:
        - name: post_id
          in: path
          description: Идентификатор публикации
          required: true
          schema:
            type: integer
        - name: component_id
          in: path
          description: Идентификатор компонента
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Сведения о компоненте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Component'
        422:
          description: Невалидные данные
          content: { }
        500:
          description: Внутренние ошибки
          content: { }
      x-codegen-request-body-name: body
    put:
      tags:
        - components
      summary: Обновить компонент
      operationId: updateComponent
      requestBody:
        description: Актуальные сведения о компоненте
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Component'
        required: true
      parameters:
        - name: post_id
          in: path
          description: Идентификатор публикации
          required: true
          schema:
            type: integer
        - name: component_id
          in: path
          description: Идентификатор компонента
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Обновлённый компонент
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Component'
        422:
          description: Невалидные данные
          content: { }
        500:
          description: Внутренние ошибки
          content: { }
      x-codegen-request-body-name: body
    delete:
      tags:
        - components
      summary: Удалить компонент
      operationId: deleteComponent
      parameters:
        - name: post_id
          in: path
          description: Идентификатор публикации
          required: true
          schema:
            type: integer
        - name: component_id
          in: path
          description: Идентификатор компонента
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Удаление завершено успешно
          content: { }
        422:
          description: Невалидные данные
          content: { }
        500:
          description: Внутренние ошибки
          content: { }
      x-codegen-request-body-name: body
components:
  schemas:
    NewPost:
      type: object
      properties:
        name:
          type: string
          description: Внутреннее название публикации
        seo_url:
          type: string
          description: Человекочитаемый URL
    Post:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          description: Внутреннее название публикации
        seo_url:
          type: string
          description: Человекочитаемый URL
    PostReference:
      type: object
      properties:
        post_id:
          type: integer
          description: Идентификатор публикации
    NewComponent:
      type: object
      properties:
        public:
          type: boolean
          description: Является ли компонент публичным
        type:
          type: string
          enum: [ string, boolean, relation ]
          description: Тип значения компонента (строка, булево, вложения)
        value:
          oneOf:
            - type: string
            - type: boolean
            - type: array
              items:
                $ref: '#/components/schemas/PostReference'
          description: Значение компонента
        order:
          type: integer
          description: Порядковый номер компонента в публикации (0=первый)
        display_class:
          type: string
          description: Класс для отображения на клиентской стороне
        custom_fields:
          type: object
          additionalProperties:
            type: string
          description: Настраиваемые поля (словарь)
    Component:
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор компонента
        post_id:
          type: integer
          description: Идентификатор публикации
        public:
          type: boolean
          description: Является ли компонент публичным
        type:
          type: string
          enum: [ string, boolean, relation ]
          description: Тип значения компонента (строка, булево, вложения)
        value:
          oneOf:
            - type: string
            - type: boolean
            - type: array
              items:
                $ref: '#/components/schemas/PostReference'
          description: Значение компонента
        order:
          type: integer
          description: Порядковый номер компонента в публикации (0=первый)
        display_class:
          type: string
          description: Класс для отображения на клиентской стороне
        custom_fields:
          type: object
          additionalProperties:
            type: string
          description: Настраиваемые поля (словарь)